+++ src/main/java/com/book/service/LoginService.java
@@ -2,7 +2,7 @@ package com.book.service;
 
 public interface LoginService {
 	
-	boolean isAdmin(Integer id,String passwd);
+	boolean isAdmin(Long id,String passwd);
 	
-	boolean isReader(Integer id,String passwd);
+	boolean isReader(Long id,String passwd);
 }
